Extract isAnnotatedFunctionPropsDeclaration for reuse
